move hex functions to internal
